Move getMaxMonth to cal utils
